requirements.txt
ipykernel 
jupyter
python-dotenv


llama_stack_client==0.2.9
streamlit
fire
